requirements.txt
# =======================
# File: requirements.txt
# =======================
watchdog
openpyxl
python-dotenv
requests